Check Exploration Algorithm logic ( not working)
Remember that door is  88 cm wide and the robot radius is 25. Now, the algorithm doesn't work because the vector distances are zero.. Then I have to simulate in Gazebo to obtain a global/local costmap that are wide enough. Therefore tweak all the parameters and test it on real robot.

REVIEW ALL THE LOGIC OF THE CODE AND PERFORM AN INTENSIVE REFACTOR.